+++ library/src/main/java/com/davidstemmer/screenplay/scene/component/CallbackComponent.java
@@ -1,7 +1,5 @@
 package com.davidstemmer.screenplay.scene.component;
 
-import android.content.Context;
-
 import com.davidstemmer.screenplay.scene.Scene;
 
 /**
@@ -18,11 +16,11 @@ public class CallbackComponent<R> implements Scene.Component {
     }
 
     @Override
-    public void afterSetUp(Context context, Scene scene) {}
+    public void afterSetUp(Scene scene, boolean isInitializing) {}
 
     @Override
-    public void beforeTearDown(Context context, Scene scene, boolean isSceneFinishing) {
-        if (isSceneFinishing) {
+    public void beforeTearDown(Scene scene, boolean isFinishing) {
+        if (isFinishing) {
             callback.onExitScene(resultHandler.getResult());
             resultHandler.reset();
         }